requirements.txt
customtkinter==5.2.2
mediapipe==0.10.14
numpy==1.26.4
opencv-python==4.10.0.82
pandas==2.2.2
Pillow==10.4.0
pyttsx3==2.90
scikit-learn==1.5.1
tensorflow==2.16.2